Replace index.js blit flow with WindowVm submitAnimationFrame flow
