+++ src/index.js
@@ -47,7 +47,7 @@ const {version} = require('../package.json');
 const symbols = require('./symbols');
 const THREE = require('../lib/three-min.js');
 const nativeBindings = require('./native-bindings.js');
-const {tickAnimationFrame, updateXrFrame, postEvent} = require('./WindowVm.js');
+const {tickAnimationFrame, submitAnimationFrame, updateXrFrame, postEvent} = require('./WindowVm.js');
 const {VRFrameData, VRStageParameters, getAllGamepads} = require('./VR.js');
 const {nativeVideo, nativeVr, nativeLm, nativeMl, nativeWindow, nativeAnalytics} = nativeBindings;
 const vmOne = require('vm-one');
@@ -349,71 +349,6 @@ const _bindWindow = (window, newWindowCb) => {
     console.dir({width, height, image: name, result: result.length});
     fs.writeFileSync(name, result);
   }; */
-  const _blit = () => {
-    for (let i = 0; i < contexts.length; i++) {
-      const context = contexts[i];
-
-      const isDirty = (!!context.isDirty && context.isDirty()) || mlPresentState.mlGlContext === context;
-      if (isDirty) {
-        const windowHandle = context.getWindowHandle();
-        
-        nativeWindow.setCurrentWindowContext(windowHandle);
-        if (isMac) {
-          context.flush();
-        }
-
-        const isVisible = nativeWindow.isVisible(windowHandle) || vrPresentState.glContext === context || mlPresentState.mlGlContext === context;
-        if (isVisible) {
-          if (vrPresentState.glContext === context && vrPresentState.hasPose) {
-            if (vrPresentState.layers.length > 0) {
-              nativeWindow.composeLayers(context, vrPresentState.fbo, vrPresentState.layers);
-            } else {
-              nativeWindow.blitFrameBuffer(context, vrPresentState.msFbo, vrPresentState.fbo, vrPresentState.glContext.canvas.width, vrPresentState.glContext.canvas.height, vrPresentState.glContext.canvas.width, vrPresentState.glContext.canvas.height, true, false, false);
-            }
-
-            vrPresentState.compositor.Submit(context, vrPresentState.tex);
-            vrPresentState.hasPose = false;
-
-            nativeWindow.blitFrameBuffer(context, vrPresentState.fbo, 0, vrPresentState.glContext.canvas.width * (args.blit ? 0.5 : 1), vrPresentState.glContext.canvas.height, window.innerWidth, window.innerHeight, true, false, false);
-          } else if (mlPresentState.mlGlContext === context && mlPresentState.mlHasPose) {
-            if (mlPresentState.layers.length > 0) { // TODO: composition can be directly to the output texture array
-              nativeWindow.composeLayers(context, mlPresentState.mlFbo, mlPresentState.layers);
-            } else {
-              nativeWindow.blitFrameBuffer(context, mlPresentState.mlMsFbo, mlPresentState.mlFbo, mlPresentState.mlGlContext.canvas.width, mlPresentState.mlGlContext.canvas.height, mlPresentState.mlGlContext.canvas.width, mlPresentState.mlGlContext.canvas.height, true, false, false);
-            }
-
-            mlPresentState.mlContext.SubmitFrame(context, mlPresentState.mlFbo, mlPresentState.mlGlContext.canvas.width, mlPresentState.mlGlContext.canvas.height);
-            mlPresentState.mlHasPose = false;
-
-            // nativeWindow.blitFrameBuffer(context, mlPresentState.mlFbo, 0, mlPresentState.mlGlContext.canvas.width, mlPresentState.mlGlContext.canvas.height, window.innerWidth, window.innerHeight, true, false, false);
-          } else if (fakePresentState.layers.length > 0) {
-            nativeWindow.composeLayers(context, 0, fakePresentState.layers);
-          }
-        }
-
-        if (isMac) {
-          context.bindFramebufferRaw(context.FRAMEBUFFER, null);
-        }
-        nativeWindow.swapBuffers(windowHandle);
-        if (isMac) {
-          const drawFramebuffer = context.getFramebuffer(context.DRAW_FRAMEBUFFER);
-          if (drawFramebuffer) {
-            context.bindFramebuffer(context.DRAW_FRAMEBUFFER, drawFramebuffer);
-          }
-
-          const readFramebuffer = context.getFramebuffer(context.READ_FRAMEBUFFER);
-          if (readFramebuffer) {
-            context.bindFramebuffer(context.READ_FRAMEBUFFER, readFramebuffer);
-          }
-        }
-
-        /* numDirtyFrames++;
-        _checkDirtyFrameTimeout(); */
-
-        context.clearDirty();
-      }
-    }
-  }
 
   let lastFrameTime = Date.now();
   const timestamps = {
@@ -838,7 +773,7 @@ const _bindWindow = (window, newWindowCb) => {
       _saveImage(contexts[contexts.length - 1], args.image);
       process.exit(0);
     } */
-    _blit();
+    submitAnimationFrame();
     if (args.performance) {
       const now = Date.now();
       const diff = now - timestamps.last;
